+++ demo/anti_flood_ftpd.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2007-2016 Giampaolo Rodola' <g.rodola@gmail.com>.
+# Copyright (C) 2007 Giampaolo Rodola' <g.rodola@gmail.com>.
 # Use of this source code is governed by MIT license that can be
 # found in the LICENSE file.
 
@@ -72,12 +72,13 @@ class AntiFloodHandler(FTPHandler):
 
 def main():
     authorizer = DummyAuthorizer()
-    authorizer.add_user('user', '12345', '.', perm='elradfmw')
+    authorizer.add_user('user', '12345', '.', perm='elradfmwMT')
     authorizer.add_anonymous('.')
     handler = AntiFloodHandler
     handler.authorizer = authorizer
     server = FTPServer(('', 2121), handler)
     server.serve_forever(timeout=1)
+
 
 if __name__ == '__main__':
     main()