+++ demo/basic_ftpd.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2007-2016 Giampaolo Rodola' <g.rodola@gmail.com>.
+# Copyright (C) 2007 Giampaolo Rodola' <g.rodola@gmail.com>.
 # Use of this source code is governed by MIT license that can be
 # found in the LICENSE file.
 
 """A basic FTP server which uses a DummyAuthorizer for managing 'virtual
-users', setting a limit for incoming connections.
+users', setting a limit for incoming connections and a range of passive
+ports.
 """
 
 import os
@@ -21,7 +22,7 @@ def main():
 
     # Define a new user having full r/w permissions and a read-only
     # anonymous user
-    authorizer.add_user('user', '12345', os.getcwd(), perm='elradfmwM')
+    authorizer.add_user('user', '12345', os.getcwd(), perm='elradfmwMT')
     authorizer.add_anonymous(os.getcwd())
 
     # Instantiate FTP handler class
@@ -34,7 +35,7 @@ def main():
     # Specify a masquerade address and the range of ports to use for
     # passive connections.  Decomment in case you're behind a NAT.
     # handler.masquerade_address = '151.25.42.11'
-    # handler.passive_ports = range(60000, 65535)
+    handler.passive_ports = range(60000, 65535)
 
     # Instantiate FTP server class and listen on 0.0.0.0:2121
     address = ('', 2121)
@@ -46,6 +47,7 @@ def main():
 
     # start ftp server
     server.serve_forever()
+
 
 if __name__ == '__main__':
     main()