+++ demo/md5_ftpd.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2007-2016 Giampaolo Rodola' <g.rodola@gmail.com>.
+# Copyright (C) 2007 Giampaolo Rodola' <g.rodola@gmail.com>.
 # Use of this source code is governed by MIT license that can be
 # found in the LICENSE file.
 
@@ -35,12 +35,13 @@ def main():
     # get a hash digest from a clear-text password
     hash = md5(b'12345').hexdigest()
     authorizer = DummyMD5Authorizer()
-    authorizer.add_user('user', hash, os.getcwd(), perm='elradfmw')
+    authorizer.add_user('user', hash, os.getcwd(), perm='elradfmwMT')
     authorizer.add_anonymous(os.getcwd())
     handler = FTPHandler
     handler.authorizer = authorizer
     server = FTPServer(('', 2121), handler)
     server.serve_forever()
+
 
 if __name__ == "__main__":
     main()