+++ demo/tls_ftpd.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2007-2016 Giampaolo Rodola' <g.rodola@gmail.com>.
+# Copyright (C) 2007 Giampaolo Rodola' <g.rodola@gmail.com>.
 # Use of this source code is governed by MIT license that can be
 # found in the LICENSE file.
 
@@ -22,7 +22,7 @@ CERTFILE = os.path.abspath(os.path.join(os.path.dirname(__file__),
 
 def main():
     authorizer = DummyAuthorizer()
-    authorizer.add_user('user', '12345', '.', perm='elradfmw')
+    authorizer.add_user('user', '12345', '.', perm='elradfmwMT')
     authorizer.add_anonymous('.')
     handler = TLS_FTPHandler
     handler.certfile = CERTFILE
@@ -32,6 +32,7 @@ def main():
     # handler.tls_data_required = True
     server = FTPServer(('', 2121), handler)
     server.serve_forever()
+
 
 if __name__ == '__main__':
     main()