+++ pyftpdlib/test/runner.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2007-2016 Giampaolo Rodola' <g.rodola@gmail.com>.
+# Copyright (C) 2007 Giampaolo Rodola' <g.rodola@gmail.com>.
 # Use of this source code is governed by MIT license that can be
 # found in the LICENSE file.
 
@@ -12,7 +12,6 @@ from pyftpdlib.test import remove_test_files
 from pyftpdlib.test import unittest
 from pyftpdlib.test import VERBOSITY
 
-
 HERE = os.path.abspath(os.path.dirname(__file__))
 
 
@@ -23,6 +22,8 @@ def main():
     remove_test_files()
     suite = unittest.TestSuite()
     for t in testmodules:
+        # ...so that "make test" will print the full test paths
+        t = "pyftpdlib.test.%s" % t
         suite.addTest(unittest.defaultTestLoader.loadTestsFromName(t))
     result = unittest.TextTestRunner(verbosity=VERBOSITY).run(suite)
     return result.wasSuccessful()